Fixing merge chances.

+++ include/field_effect.h
@@ -41,7 +41,7 @@ void StartEscalatorWarp(u8 metatileBehavior, u8 priority);
 void StartLavaridgeGymB1FWarp(u8 priority);
 void StartLavaridgeGym1FWarp(u8 priority);
 
-u8 CountPartyNonEggMons(void);
+u32 CountPartyNonEggMons(void);
 
 void SpriteCB_AshPuff(struct Sprite *);
 void SpriteCB_AshLaunch(struct Sprite *);
